+++ core/amber/src/test/scala/edu/uci/ics/amber/engine/e2e/TestUtils.scala
@@ -19,9 +19,24 @@
 
 package edu.uci.ics.amber.engine.e2e
 
+import edu.uci.ics.amber.config.StorageConfig
 import edu.uci.ics.amber.core.workflow.WorkflowContext
 import edu.uci.ics.amber.engine.architecture.controller.Workflow
 import edu.uci.ics.amber.operator.LogicalOp
+import edu.uci.ics.texera.dao.SqlServer
+import edu.uci.ics.texera.dao.jooq.generated.enums.UserRoleEnum
+import edu.uci.ics.texera.dao.jooq.generated.tables.daos.{
+  UserDao,
+  WorkflowDao,
+  WorkflowExecutionsDao,
+  WorkflowVersionDao
+}
+import edu.uci.ics.texera.dao.jooq.generated.tables.pojos.{
+  User,
+  WorkflowExecutions,
+  WorkflowVersion,
+  Workflow => WorkflowPojo
+}
 import edu.uci.ics.texera.web.model.websocket.request.LogicalPlanPojo
 import edu.uci.ics.texera.workflow.{LogicalLink, WorkflowCompiler}
 
@@ -38,6 +53,80 @@ object TestUtils {
     workflowCompiler.compile(
       LogicalPlanPojo(operators, links, List(), List())
     )
+  }
+
+  /**
+    * If a test case accesses the user system through singleton resources that cache the DSLContext (e.g., executes a
+    * workflow, which accesses WorkflowExecutionsResource), we use a separate texera_db specifically for such test cases.
+    * Note such test cases need to clean up the database at the end of running each test case.
+    */
+  def initiateTexeraDBForTestCases(): Unit = {
+    SqlServer.initConnection(
+      StorageConfig.jdbcUrlForTestCases,
+      StorageConfig.jdbcUsername,
+      StorageConfig.jdbcPassword
+    )
+  }
+
+  val testUser: User = {
+    val user = new User
+    user.setUid(Integer.valueOf(1))
+    user.setName("test_user")
+    user.setRole(UserRoleEnum.ADMIN)
+    user.setPassword("123")
+    user.setEmail("test_user@test.com")
+    user
+  }
+
+  val testWorkflowEntry: WorkflowPojo = {
+    val workflow = new WorkflowPojo
+    workflow.setName("test workflow")
+    workflow.setWid(Integer.valueOf(1))
+    workflow.setContent("test workflow content")
+    workflow.setDescription("test description")
+    workflow
+  }
+
+  val testWorkflowVersionEntry: WorkflowVersion = {
+    val workflowVersion = new WorkflowVersion
+    workflowVersion.setWid(Integer.valueOf(1))
+    workflowVersion.setVid(Integer.valueOf(1))
+    workflowVersion.setContent("test version content")
+    workflowVersion
+  }
+
+  val testWorkflowExecutionEntry: WorkflowExecutions = {
+    val workflowExecution = new WorkflowExecutions
+    workflowExecution.setEid(Integer.valueOf(1))
+    workflowExecution.setVid(Integer.valueOf(1))
+    workflowExecution.setUid(Integer.valueOf(1))
+    workflowExecution.setStatus(3.toByte)
+    workflowExecution.setEnvironmentVersion("test engine")
+    workflowExecution
+  }
+
+  def setUpWorkflowExecutionData(): Unit = {
+    val dslConfig = SqlServer.getInstance().context.configuration()
+    val userDao = new UserDao(dslConfig)
+    val workflowDao = new WorkflowDao(dslConfig)
+    val workflowExecutionsDao = new WorkflowExecutionsDao(dslConfig)
+    val workflowVersionDao = new WorkflowVersionDao(dslConfig)
+    userDao.insert(testUser)
+    workflowDao.insert(testWorkflowEntry)
+    workflowVersionDao.insert(testWorkflowVersionEntry)
+    workflowExecutionsDao.insert(testWorkflowExecutionEntry)
+  }
+
+  def cleanupWorkflowExecutionData(): Unit = {
+    val dslConfig = SqlServer.getInstance().context.configuration()
+    val userDao = new UserDao(dslConfig)
+    val workflowDao = new WorkflowDao(dslConfig)
+    val workflowExecutionsDao = new WorkflowExecutionsDao(dslConfig)
+    val workflowVersionDao = new WorkflowVersionDao(dslConfig)
+    workflowExecutionsDao.deleteById(1)
+    workflowVersionDao.deleteById(1)
+    workflowDao.deleteById(1)
+    userDao.deleteById(1)
   }
 
 }
